initial work on background data loader
use a celluloid actors that can load data in the background to free up the main worfklow instances from running the data loads. This enable threading web servers (puma) to not block waiting for a workflow to load data. This blocking explains why heavy loading events in cellect cause it to fail the load balancer health checks and thus fall over.