respect group_id when selecting data
Some subject sets are not available as they may be all retired / empty. Now we should not randomly select from a group when asking for specific group data. Instead, set an empty set_klass for that group and sample from this empty set. If no group is passed in then randomly select data.